+++ src/libs/TokenizerSaveLocations.js
@@ -2,7 +2,11 @@ import logger from "../libs/logger.js";
 import CONSTANTS from "../constants.js";
 import DirectoryPicker from "./DirectoryPicker.js";
 
-export class TokenizerSaveLocations extends FormApplication {
+const { ApplicationV2, HandlebarsApplicationMixin } = foundry.applications.api;
+const FPClass = foundry.applications?.apps?.FilePicker?.implementation ?? FilePicker;
+
+
+export class TokenizerSaveLocations extends HandlebarsApplicationMixin(ApplicationV2) {
 
   constructor(tokenizer) {
     super();
@@ -10,40 +14,45 @@ export class TokenizerSaveLocations extends FormApplication {
     this.data = [];
   }
 
+  static PARTS = {
+    form: {
+      template: "modules/vtta-tokenizer/templates/save-locations/form.hbs",
+    },
+    footer: {
+      template: "modules/vtta-tokenizer/templates/save-locations/footer.hbs",
+    },
+  };
 
-  static get defaultOptions() {
-    const options = super.defaultOptions;
-    options.id = "tokenizer-save-locations";
-    options.template = "modules/vtta-tokenizer/templates/file-paths.hbs";
-    options.width = 500;
-    return options;
-  }
+    /** @inheritDoc */
+  static DEFAULT_OPTIONS = {
+    id: "tokenizer-save-locations",
+    classes: ["standard-form"],
+    actions: {
+      selectDirectory: TokenizerSaveLocations.selectDirectory,
+    },
+    position: {
+      width: "500",
+      height: "auto",
+    },
+    window: {
+      resizable: false,
+      minimizable: false,
+      subtitle: "",
+    },
+    tag: "form",
+    form: {
+      handler: TokenizerSaveLocations.formHandler,
+      submitOnChange: false,
+      closeOnSubmit: false,
+    },
+  };
 
   // eslint-disable-next-line class-methods-use-this
   get title() {
     return game.i18n.localize("vtta-tokenizer.label.save-locations");
   }
 
-  // in foundry v10 we no longer get read only form elements back
-  /** @override */
-  _getSubmitData(updateData = {}) {
-    let data = super._getSubmitData(updateData);
-
-    for (const element of this.form.elements) {
-      if (element.readOnly) {
-        const name = element.name;
-        const field = this.form.elements[name];
-        foundry.utils.setProperty(data, name, field.value);
-      }
-    }
-
-    return data;
-  }
-
-  /** @override */
-  // eslint-disable-next-line class-methods-use-this
-  async getData() {
-
+  async _prepareContext() {
     this.data = [
       {
         key: "avatar",
@@ -62,15 +71,51 @@ export class TokenizerSaveLocations extends FormApplication {
     return { type: this.data };
   }
 
-  /** @override */
+  static async selectDirectory(event, target) {
+    const targetDirSetting = target.dataset.target;
+    const currentDir = this.tokenizer[targetDirSetting];
+    // const parsedDir = FileHelper.parseDirectory(currentDir);
+    const current = await DirectoryPicker.getFileUrl(currentDir, "");
 
-  async _updateObject(event, formData) {
-    event.preventDefault();
+    const filePicker = new FPClass({
+      type: "folder",
+      current: current,
+      // source: parsedDir.activeSource,
+      // activeSource: parsedDir.activeSource,
+      // bucket: parsedDir.bucket,
+      callback: async (path, picker) => {
+        const activeSource = picker.activeSource;
+        const bucket = activeSource === "s3" && picker.sources.s3?.bucket && picker.sources.s3.bucket !== ""
+          ? picker.sources.s3.bucket
+          : null;
 
+        const formattedPath = DirectoryPicker.format({
+          activeSource,
+          bucket,
+          path,
+        });
+
+        this.element.querySelector(`input[name='${targetDirSetting}']`).value = formattedPath;
+      },
+    });
+    filePicker.render();
+
+  }
+
+
+  /**
+   * Process form submission for the sheet
+   * @this {DDBLocationSetup}                      The handler is called with the application as its bound scope
+   * @param {SubmitEvent} event                   The originating form submission event
+   * @param {HTMLFormElement} form                The form element that was submitted
+   * @param {FormDataExtended} formData           Processed data for the submitted form
+   * @returns {Promise<void>}
+   */
+  static async formHandler(event, form, formData) {
     const directoryStatus = [];
 
     for (const dataType of this.data) {
-      const value = formData[`${dataType.key}UploadDirectory`];
+      const value = formData.object[`${dataType.key}UploadDirectory`];
       // eslint-disable-next-line no-await-in-loop
       directoryStatus.push({
         key: dataType.key,
@@ -82,38 +127,40 @@ export class TokenizerSaveLocations extends FormApplication {
     }
 
     if (directoryStatus.some((dir) => dir.isBad)) {
-      $("tokenizer-directory-setup").text(
-        `Please set the image upload directory(s) to something other than the root.`,
-      );
-      $("#ddb-importer-folders").css("height", "auto");
-      logger.error("Error setting Image directory", {
-        directoryStatus,
-      });
-      throw new Error(
-        `Please set the image upload directory to something other than the root.`,
-      );
+      for (const data of directoryStatus.filter((dir) => dir.isBad)) {
+        ui.notifications.error(
+          `Please set the image upload directory for ${data.name} to something other than the root.`,
+          { permanent: true },
+        );
+        logger.error("Error setting Image directory", {
+          directoryStatus,
+          data,
+        });
+      }
     } else if (directoryStatus.some((dir) => !dir.isValid)) {
-      $("#munching-folder-setup").text(`Directory Validation Failed.`);
-      $("#ddb-importer-folders").css("height", "auto");
-      logger.error("Error validating Image directory", {
-        directoryStatus,
-      });
-      throw new Error(`Directory Validation Failed.`);
+      for (const data of directoryStatus.filter((dir) => !dir.isValid)) {
+        ui.notifications.error(
+          `Directory Validation Failed for ${data.name} please check it exists and can be written to.`,
+          { permanent: true },
+        );
+        logger.error("Error validating Image directory", {
+          directoryStatus,
+          data,
+        });
+      }
     } else {
-      this.tokenizer.avatarUploadDirectory = formData["avatarUploadDirectory"];
-      this.tokenizer.tokenUploadDirectory = formData["tokenUploadDirectory"];
-      this.tokenizer.avatarFileName = formData["avatarFileName"];
-      this.tokenizer.tokenFileName = formData["tokenFileName"];
+      this.tokenizer.avatarUploadDirectory = formData.object["avatarUploadDirectory"];
+      this.tokenizer.tokenUploadDirectory = formData.object["tokenUploadDirectory"];
+      this.tokenizer.avatarFileName = formData.object["avatarFileName"];
+      this.tokenizer.tokenFileName = formData.object["tokenFileName"];
       logger.debug("Changed tokenizer save paths to...", {
         avatarUploadDirectory: this.tokenizer.avatarUploadDirectory,
         tokenUploadDirectory: this.tokenizer.tokenUploadDirectory,
         avatarFileName: this.tokenizer.avatarFileName,
         tokenFileName: this.tokenizer.tokenFileName,
       });
+      this.close();
     }
+
   }
 }
-
-Hooks.on("renderTokenizerSaveLocations", (app, html) => {
-  DirectoryPicker.processHtml(html);
-});